Rectify division by 0 fix
Zero length edges on the UVs have been generating errors if were part of the active face. The best solution is not having them, or selecting the least distorted, non zero face as active before using Rectify. This only helps avoiding the error to arise. Fixes #77
No more loading of unused modules and classes on Rectify.