Bake color space
Added ability to choose the output color space for the texture per bake, a much needed option to keep coherence in pbr workflows for channels as roughness, especilly when using the generated maps in other programs/environments. An option in the preferences could come later to automate the process of choosing linear or srgb depending in the bake mode.