Fixed variable name collision with built-in "set"
Changed variable name "set" to "bset" so it wouldn't clobber Python's
built-in "set" function. set() is used many times in other places in
the codebase as well so this should prevent future problems in those scopes.
Also changed 2 uses of range(0,len()) to enumerate() as its more
pythonic, more terse, and does the same thing.