VC limit reached warning
Some bake modes use internally a custom Vertex Color layer which is removed once the bake finishes. If the maximum number of VC layers are in use in an object to be baked from (that is, 8), warn about it and gray out the Bake button to avoid an exception.

Additionally, make sure materials without node tree don't make the addon UI refresh fail in the search of UDIMs, writing an error in the Console..