Texel Density to support UDIMs better while in Image Get Mode
The resolution of the image bound to the active UDIM Tile is taken into consideration when computing texel densities in both get and set modes. This follows the workflow established by the TexTools UV Layout UDIM support added months ago; that is, the user must manage the active UDIM Tile as needed in the first TexTools panel for the tools to work based on it, which would be sufficient in most common scenarios.
For some utilities, the option of getting the UDIM Tile closest to every UV island on the selection may be nice to have but would be overkill for the script to work. Anyway, it may be added in the future.