Flush selection fix
Due to changes in recent versions of Blender regarding the UV Space selection modes, flushing of the selection from UV loops to UV edges and faces is not working as before. The outcome is the lost of the selection once certain TexTools utilities are used. Added a workaround to several operators to recover the original functionality.