Point to the IDs of the Principled BSDF nodes
Some utilities of this addon work almost uniquely on Principled BSDF shader nodes. Discovering them in the node tree was relying in the node name, not in the id name, so when it is translated internally or modified by the user, some utilities were failing. Now we point to them in a proper way.
Some corrections in op_texel_checker_map to flush unused data in all scenarios.
Should fix #158