Remark new minimum Blender version requirements
Some operators have been updated to newer versions of the Blender Python API. Now, the minimum version of Blender which is fully supported is 3.2. This should be expected to change in the future to take benefit from the faster access to the new data structures of Blender 4.0 that the latest API provides.