Baking, better avoid circular dependency with saved image
Clear users of the saved image and recover them for the newly baked image (they share the ID but are not the same); this seems to solve the multi-user escalation of reiterated baking over a saved image, and allows the newly baked image to be saved over the old.
Recover detection of the usage of baked-over unsaved-images on the node-tree; it's slower than ID remapping, but enable the use of an alert when circular dependency can indeed happen during a bake.
Recovered old removal of previous or temporary images; errors prompted in the console can't be avoided in any case, either an erase of users is forced or an image with no real users is removed.