Fixed detection of active Vertex Colors layer
Since Blender 3.4, reaching the active Vertex Colors layer from bmesh is not working as before, so I changed the approach. Fixes #166
Partially Fixes #175 in regards of Vertex Colors assigning, conversions only worked with materials before as far as I remember, and other things are as consistent as before. Maybe a redesign of the area is needed. A new issue could be opened anyway to tackle those.