Expand support for single Active-but-unselected Objects
In blender. you can work in Edit Mode with an unselected object if it is still tagged as Active internally, but TexTools doesn't allow it. We have received issues related to this a number of times, and it has been solved for a number of tools, but not all. When no object is selected, the active-but-unselected Object should be made selected temporarily for the add-on to work as expected. For now, this works only for some tools which still use the multi_object_loop function. I think that the Object should remain as unselected afterwards, and not be considered in multi-selections.
Should fix #204