Added per object control for Checker Map scale
Formerly controlled with the TexTools global resolution, now a Custom Property is added to every Object which has a Checker Map applied to tune the tiling (or projection Scale) of the Checker texture. For convenience, the value of this property for the Active Object is exposed in the Textools UI below the Checker Map button.
Now, the Viewport Shading Modes "Material Preview" or "Rendered" need to be in use to view the effect of the Checker Map Tiling.