Clarify Select Flipped, Select Zero behaviour
These operators act on all geometry faces, even hidden or unselected, which is not necessarily bad, but may be confusing under some common scenarios, as no other tools behaves that way. I decided to keep them like that, so I'll just try to clarify how they work in the descriptions and warnings.
Somehow this Fixes #233